Refactor logging.py for clarity and organization
