+++ logger/__init__.py
@@ -1,0 +1,7 @@
+import logging
+from logger.log_config import init_logger
+
+log = logging.getLogger('WILL')
+
+
+init_logger(log)
